% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_path.R
\name{assocPlotRscript}
\alias{assocPlotRscript}
\title{Get path to the assoc_plot.R script}
\usage{
assocPlotRscript()
}
\value{
Character string containing the path to the assoc_plot.R script
}
\description{
Get path to the assoc_plot.R script
}
\examples{
\dontrun{
# Get path to assoc_plot.R script
assoc_plot_script <- assocPlotRscript()

 Run the script on the termninal
 requires ggplot2 and data.table
}

}
