% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFChromosomes}
\alias{BCFChromosomes}
\title{Get indexed chromosomes}
\usage{
BCFChromosomes(fp)
}
\arguments{
\item{fp}{the vcf reader}
}
\value{
a list of chromosome
}
\description{
list the indexed chromosomes
}
