% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFClose}
\alias{BCFClose}
\title{Close a VCF reader}
\usage{
BCFClose(fp)

BCFClose(fp)
}
\arguments{
\item{fp}{the vcf reader}
}
\value{
true on success

true on success
}
\description{
Close a VCF reader

fp<-BCFOpen("my.vcf.gz",TRUE)
BCFClose(fp)
}
\details{
fp<-BCFOpen("my.vcf.gz",TRUE)
BCFClose(fp)
}
\examples{
\dontrun{
fp <- BCFOpen("my.vcf.gz", TRUE)
BCFClose(fp)
}
}
