% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFFilters}
\alias{BCFFilters}
\title{Get filter table from VCF/BCF header}
\usage{
BCFFilters(fp)
}
\arguments{
\item{fp}{the vcf reader}
}
\value{
a table of filters
}
\description{
Get filter table from VCF/BCF header
}
