% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFFormats}
\alias{BCFFormats}
\title{Get FORMAT table from VCF/BCF header}
\usage{
BCFFormats(fp)
}
\arguments{
\item{fp}{the vcf reader}
}
\value{
a table of filters
}
\description{
Get FORMAT table from VCF/BCF header
}
