% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFInfos}
\alias{BCFInfos}
\title{Get INFO table from VCF/BCF header}
\usage{
BCFInfos(fp)
}
\arguments{
\item{fp}{the vcf reader}
}
\value{
a table of filters
}
\description{
Get INFO table from VCF/BCF header
}
