% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFNewWriter}
\alias{BCFNewWriter}
\title{Open a new VCF writer}
\usage{
BCFNewWriter(fp, fname = "-")
}
\arguments{
\item{fp}{the vcf reader}

\item{fname}{the name of the output vcf file}
}
\value{
the writer or NULL on failure
}
\description{
Open a new VCF writer.
Must be closed with BCFClose
}
\examples{
\dontrun{
fp <- BCFOpen("my.vcf.gz", TRUE)
out <- BCFNewWriter(fp, "output.vcf")
BCFClose(out)
}
}
