% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFNext}
\alias{BCFNext}
\title{Read next variant context}
\usage{
BCFNext(fp)
}
\arguments{
\item{fp}{the vcf reader}
}
\value{
an opaque vcf context or NULL
}
\description{
read the next Variant Context in the VCF reader
}
\examples{
\dontrun{
fp <- BCFOpen("in.vcf.gz")
BCFQuery(fp,"RF02:1-1000")
while(!is.null(vc<-BCFNext(fp))) {
     ## do something with vc
     }
BCFClose(fp)
}
}
