% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFNSamples}
\alias{BCFNSamples}
\title{Get number of samples}
\usage{
BCFNSamples(fp)
}
\arguments{
\item{fp}{the vcf reader}
}
\value{
the number of samples
}
\description{
Get number of samples in VCF/BCF file
}
