% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFOpen}
\alias{BCFOpen}
\title{Open a VCF or BCF file}
\usage{
BCFOpen(filename, requireIndex = TRUE)
}
\arguments{
\item{filename}{the path to the vcf file}

\item{requireIndex}{load associated vcf index}
}
\value{
the new VCF reader
}
\description{
Open a VCF or a BCF file
}
\details{
fp<-BCFOpen("my.vcf.gz",TRUE)
BCFClose(fp)
fp<-BCFOpen("my.bcf")
BCFClose(fp)
}
\examples{
\dontrun{
fp <- BCFOpen("my.vcf.gz", TRUE)
BCFClose(fp)
fp <- BCFOpen("my.bcf")
BCFClose(fp)
}
}
