% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFQuery}
\alias{BCFQuery}
\title{Query VCF by interval}
\usage{
BCFQuery(fp, interval, collect = FALSE)
}
\arguments{
\item{fp}{the vcf reader}

\item{interval}{the genomic interval to be scanned}

\item{collect}{if TRUE return a list of variants in the region or NULL on failure}
}
\value{
TRUE on success or a list of variant if collect=TRUE
}
\description{
prepare the VCF reader for a new vcf iteration over a given interval.
VCF reader must be associated and opened with a valid index.
}
