% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFSample2Index}
\alias{BCFSample2Index}
\title{Convert sample names to indices}
\usage{
BCFSample2Index(fp, sn)
}
\arguments{
\item{fp}{the vcf reader}

\item{sn}{the sample name}
}
\value{
1-based sample indices
}
\description{
Convert sample names to indices
}
