% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFSampleAt}
\alias{BCFSampleAt}
\title{Get sample name by index}
\usage{
BCFSampleAt(fp, idx)
}
\arguments{
\item{fp}{the vcf reader}

\item{idx}{the 1-based index}
}
\value{
the idx-th sample (1-based)
}
\description{
Get sample name by index
}
