% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_path.R
\name{BCFToolsCLIPath}
\alias{BCFToolsCLIPath}
\title{Get path to BCFTools CLI executables}
\usage{
BCFToolsCLIPath(cli = c("RBcftools", "BCFToolsCli"))
}
\arguments{
\item{cli}{Character string; which CLI to get the path for.
Either "RBcftools" or "BCFToolsCli". Default is "RBcftools".}
}
\value{
Character string containing the path to the requested CLI executable
}
\description{
Returns the path to the command-line interface executables provided by RBCFLib.
These CLI tools allow users to run bcftools commands from the command line
using the RBCFLib package's BCFToolsRun function.
}
\examples{
\dontrun{
# Get path to RBcftools
bcftools_cli <- BCFToolsCLIPath()

# Get path to BCFToolsCli (PascalCase version)
bcftools_pascal_cli <- BCFToolsCLIPath("BCFToolsCli")

# Use system2 to run commands
system2(bcftools_cli, c("view", "-h", vcfFile))
}

}
