% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{BCFToolsScoreVersion}
\alias{BCFToolsScoreVersion}
\title{Get bcftools score Version}
\usage{
BCFToolsScoreVersion()
}
\value{
A character string with the bcftools score version
}
\description{
Returns the version of the bcftools score plugin used by RBCFLib
'
}
