% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{BCFToolsVersion}
\alias{BCFToolsVersion}
\title{Get bcftools Version}
\usage{
BCFToolsVersion()
}
\value{
A character string with the bcftools version
}
\description{
Returns the version of the bcftools library used by RBCFLib
}
