% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{BCFWriteVariant}
\alias{BCFWriteVariant}
\title{Save a Variant in a VCF writer}
\usage{
BCFWriteVariant(fp, vc)
}
\arguments{
\item{fp}{the vcf reader}

\item{vc}{the variant}
}
\value{
true on success
}
\description{
Save a Variant in a VCF writer
}
\examples{
\dontrun{
fp <- BCFOpen("my.vcf.gz", TRUE)
out <- BCFNewWriter(fp, "output.vcf")
while (!is.null(vc <- BCFNext(fp))) {
  BCFWriteVariant(out, vc)
}
BCFClose(fp)
BCFClose(out)
}
}
