% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{CGRangesDestroy}
\alias{CGRangesDestroy}
\title{Destroy a CGRanges object}
\usage{
CGRangesDestroy(cr)
}
\arguments{
\item{cr}{External pointer to CGRanges object}
}
\value{
Invisible NULL
}
\description{
Frees the memory associated with a CGRanges object. After calling this function,
the CGRanges object should not be used anymore.
}
\note{
This function is automatically called when the R object is garbage collected,
but can be called explicitly to free memory immediately.
}
