% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{CGRangesExtractByIndex}
\alias{CGRangesExtractByIndex}
\title{Extract intervals by 1-based index from CGRanges}
\usage{
CGRangesExtractByIndex(cr, indices)
}
\description{
Extract intervals by 1-based index from CGRanges
}
