% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{CGRangesIndex}
\alias{CGRangesIndex}
\title{Index CGRanges for fast queries}
\usage{
CGRangesIndex(cr)
}
\arguments{
\item{cr}{External pointer to CGRanges object}
}
\value{
Invisible NULL
}
\description{
Builds an interval tree index for the CGRanges object to enable fast overlap queries.
This must be called after adding all intervals and before performing any queries.
}
\seealso{
\code{\link{CGRangesAdd}}, \code{\link{CGRangesOverlap}}
}
