% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{CGRangesOverlap}
\alias{CGRangesOverlap}
\title{Query CGRanges for overlapping intervals}
\usage{
CGRangesOverlap(cr, chrom, start, end)
}
\arguments{
\item{cr}{External pointer to indexed CGRanges object}

\item{chrom}{Character, chromosome name to query}

\item{start}{Integer, start position of query region (1-based)}

\item{end}{Integer, end position of query region (1-based, inclusive)}
}
\value{
Integer vector of indices (1-based) of overlapping intervals
}
\description{
Finds all intervals in the CGRanges object that overlap with the specified query region.
The CGRanges object must be indexed first using \code{\link{CGRangesIndex}}.
}
\seealso{
\code{\link{CGRangesIndex}}, \code{\link{CGRangesOverlapVec}}
}
