% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DecompressFile}
\alias{DecompressFile}
\title{Decompress a compressed file}
\usage{
DecompressFile(
  input_file,
  output_file = NULL,
  remove_input = FALSE,
  block_size = 1e+08
)
}
\arguments{
\item{input_file}{Path to the compressed file to decompress}

\item{output_file}{Path for the decompressed output file. If NULL, the input filename without extension is used}

\item{remove_input}{Logical; whether to remove the input file after successful decompression (default: FALSE)}

\item{block_size}{Size of data chunks to process at once (in bytes, default: 1e8)}
}
\value{
Path to the decompressed file
}
\description{
Decompresses various compressed file formats (.gz, .bz2, .xz) using base R functionality, borrowing
from the R.utils package for the underlying logic. This function is designed to work with
without relying on external dependencies.
ref : https://github.com/HenrikBengtsson/R.utils/blob/74def095eaa244e355d05fdf790ee6393dad1d99/R/compressFile.R#L15
}
