% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DownloadHumanReferenceGenomes}
\alias{DownloadHumanReferenceGenomes}
\title{Download human reference genomes (GRCh37 and GRCh38)}
\usage{
DownloadHumanReferenceGenomes(
  grch37_dir = file.path(Sys.getenv("HOME"), "GRCh37"),
  grch38_dir = file.path(Sys.getenv("HOME"), "GRCh38"),
  urls = list(grch37_fasta =
    "http://ftp.ncbi.nlm.nih.gov/1000genomes/ftp/technical/reference/human_g1k_v37.fasta.gz",
    grch38_fasta =
    "http://ftp.ncbi.nlm.nih.gov/genomes/all/GCA/000/001/405/GCA_000001405.15_GRCh38/seqs_for_alignment_pipelines.ucsc_ids/GCA_000001405.15_GRCh38_no_alt_analysis_set.fna.gz",
    grch37_cytoband =
    "http://hgdownload.cse.ucsc.edu/goldenPath/hg19/database/cytoBand.txt.gz",
    grch38_cytoband =
    "http://hgdownload.cse.ucsc.edu/goldenPath/hg38/database/cytoBand.txt.gz",
    grch37_chain =
    "http://hgdownload.cse.ucsc.edu/goldenpath/hg18/liftOver/hg18ToHg19.over.chain.gz", 

        grch38_chain_18 =
    "http://hgdownload.cse.ucsc.edu/goldenpath/hg18/liftOver/hg18ToHg38.over.chain.gz",
    grch38_chain_19 =
    "http://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/hg19ToHg38.over.chain.gz"),
  cytoband = FALSE,
  chain = FALSE,
  method = "wget",
  extra = getOption("download.file.extra")
)
}
\arguments{
\item{grch37_dir}{Directory to store GRCh37 reference (default: ~/GRCh37)}

\item{grch38_dir}{Directory to store GRCh38 reference (default: ~/GRCh38)}

\item{urls}{Named list of URLs for genome resources (with defaults)}

\item{cytoband}{Logical; also download cytoband files (default: FALSE)}

\item{chain}{Logical; also download chain files for liftover (default: FALSE)}

\item{method}{Method to be used for downloading files. Possible values are:
\itemize{
\item \code{"wget"} (default) - Uses the external wget utility which provides better functionality
for large files including automatic retries and the ability to resume interrupted downloads
\item \code{"curl"} - Uses the external curl utility
\item \code{"auto"} - Let R choose the best method available
\item \code{"internal"} - Uses R's internal download method
\item \code{"libcurl"} - Uses the libcurl library
\item \code{"lynx"} - Uses the lynx browser in text mode
}}

\item{extra}{Character vector of additional command-line arguments for the 'wget' and 'curl' methods.
For example, to use wget's continue feature, you can provide \code{extra = c("-C", "on")} to resume
partially downloaded files. See \code{?download.file} for more details.

The \code{wget} method is recommended for large genome files as it has excellent support for
resuming interrupted downloads, which is important when downloading multi-gigabyte files.
See \code{?download.file} for more details on these methods.}
}
\value{
Named list with paths to downloaded files
}
\description{
Downloads the GRCh37 and GRCh38 human reference FASTA files and their indexes to the specified directories.
}
