% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FaidxIndexFasta}
\alias{FaidxIndexFasta}
\title{Index a FASTA file using htslib faidx}
\usage{
FaidxIndexFasta(fasta_path)
}
\arguments{
\item{fasta_path}{Path to the FASTA file}
}
\value{
Path to the generated .fai index file
}
\description{
Index a FASTA file using htslib faidx
}
