% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeAllelesIdx0}
\alias{GenotypeAllelesIdx0}
\title{Get 0-based alleles indexes for genotype}
\usage{
GenotypeAllelesIdx0(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
max ploidy
}
\description{
return the 0-based alleles indexes for the given genotype
}
