% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeFiltered}
\alias{GenotypeFiltered}
\title{Check if genotype is filtered}
\usage{
GenotypeFiltered(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if genotype is filtered (FORMAT/FT is set)
}
\description{
Check if genotype is filtered
}
