% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeHasAd}
\alias{GenotypeHasAd}
\title{Check if AD is available for genotype}
\usage{
GenotypeHasAd(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if AD is available
}
\description{
Check if AD is available for genotype
}
