% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeHasDp}
\alias{GenotypeHasDp}
\title{Check if genotype has read depth}
\usage{
GenotypeHasDp(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if DP is available
}
\description{
Check if genotype has read depth
}
