% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeHasGq}
\alias{GenotypeHasGq}
\title{Check if genotype has quality score}
\usage{
GenotypeHasGq(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if GQ is available
}
\description{
Check if genotype has quality score
}
