% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeHasPl}
\alias{GenotypeHasPl}
\title{Check if PL is available for genotype}
\usage{
GenotypeHasPl(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if PL is available
}
\description{
Check if PL is available for genotype
}
