% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeHet}
\alias{GenotypeHet}
\title{Check if genotype is heterozygous}
\usage{
GenotypeHet(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if genotype is diploid and heterozygous
}
\description{
Check if genotype is heterozygous
}
