% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeHetNonRef}
\alias{GenotypeHetNonRef}
\title{Check if genotype is heterozygous non-reference}
\usage{
GenotypeHetNonRef(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if genotype is diploid and heterozygous and doesn't contains the alt allele
}
\description{
Check if genotype is heterozygous non-reference
}
