% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeHomRef}
\alias{GenotypeHomRef}
\title{Check if genotype is homozygous reference}
\usage{
GenotypeHomRef(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if genotype is diploid and all alleles are reference
}
\description{
Check if genotype is homozygous reference
}
