% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeHomVar}
\alias{GenotypeHomVar}
\title{Check if genotype is homozygous variant}
\usage{
GenotypeHomVar(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if genotype is diploid and homozygous on alt allele
}
\description{
Check if genotype is homozygous variant
}
