% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeIntAttribute}
\alias{GenotypeIntAttribute}
\title{Get integer attribute from genotype}
\usage{
GenotypeIntAttribute(gt, att)
}
\arguments{
\item{gt}{the genotype}

\item{att}{the key}
}
\value{
the values for this key+genotype
}
\description{
Get integer attribute from genotype
}
