% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeNoCall}
\alias{GenotypeNoCall}
\title{Check if genotype is no-call}
\usage{
GenotypeNoCall(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if genotypes contains no allele '' or any is no call '.'
}
\description{
Check if genotype is no-call
}
