% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypePhased}
\alias{GenotypePhased}
\title{Check if genotype is phased}
\usage{
GenotypePhased(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
TRUE if genotype is phased
}
\description{
Check if genotype is phased
}
