% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypePl}
\alias{GenotypePl}
\title{Get PL for genotype}
\usage{
GenotypePl(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
PL or empty vector
}
\description{
Get PL (Phred-scaled likelihoods) for genotype
}
