% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypePloidy}
\alias{GenotypePloidy}
\title{Get ploidy of genotype}
\usage{
GenotypePloidy(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
the number of alleles for the genotypes
}
\description{
Get ploidy of genotype
}
