% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeSample}
\alias{GenotypeSample}
\title{Get sample name for genotype}
\usage{
GenotypeSample(gt)
}
\arguments{
\item{gt}{the genotype}
}
\value{
the name of the sample associated to the genotype
}
\description{
Get sample name for genotype
}
