% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{GenotypeStringAttribute}
\alias{GenotypeStringAttribute}
\title{Get string attribute from genotype}
\usage{
GenotypeStringAttribute(gt, att)
}
\arguments{
\item{gt}{the genotype}

\item{att}{the key}
}
\value{
a String for this flag+genotype or NULL
}
\description{
Get string attribute from genotype
}
