% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{HTSLibVersion}
\alias{HTSLibVersion}
\title{Get htslib Version}
\usage{
HTSLibVersion()
}
\value{
A character string with the htslib version
}
\description{
Returns the version of the htslib library used by RBCFLib
}
