% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{looks_like_gt_context}
\alias{looks_like_gt_context}
\title{Check if object is a genotype context}
\usage{
looks_like_gt_context(gt)
}
\arguments{
\item{gt}{the object to check}
}
\value{
TRUE if object is a genotype context
}
\description{
Check if an object looks like a genotype context
}
