% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{looks_like_variant_context}
\alias{looks_like_variant_context}
\title{Check if object is a variant context}
\usage{
looks_like_variant_context(vc)
}
\arguments{
\item{vc}{the object to check}
}
\value{
TRUE if object is a variant context
}
\description{
Check if an object looks like a variant context
}
