% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{looks_like_vcf_context}
\alias{looks_like_vcf_context}
\title{Check if object is a VCF context}
\usage{
looks_like_vcf_context(vcf)
}
\arguments{
\item{vcf}{the object to check}
}
\value{
TRUE if object is a VCF context
}
\description{
Check if an object looks like a VCF context
}
