% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCFToolsMunge.R
\name{MungeSumstatsHeadersFile}
\alias{MungeSumstatsHeadersFile}
\title{Get the path to the sumstats headers file
as defined by MungeSumstats}
\usage{
MungeSumstatsHeadersFile()
}
\value{
Path to the sumstats headers file
}
\description{
Get the path to the sumstats headers file
as defined by MungeSumstats
}
