% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantAlleles}
\alias{VariantAlleles}
\title{Get alleles for variant}
\usage{
VariantAlleles(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the alleles for this variant
}
\description{
get the alleles for this variant
}
