% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantAltAlleles}
\alias{VariantAltAlleles}
\title{Get alternate alleles for variant}
\usage{
VariantAltAlleles(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the alternate alleles for this variant
}
\description{
get the alternate alleles for this variant
}
