% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantContig}
\alias{VariantContig}
\title{Get chromosome name for variant}
\usage{
VariantContig(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the chromosome name
}
\description{
get chromosome name for this variant
}
