% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantEnd}
\alias{VariantEnd}
\title{Get variant end position}
\usage{
VariantEnd(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the END position for this variant
}
\description{
get chromosome END for this variant
}
