% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantFilters}
\alias{VariantFilters}
\title{Get filters for variant}
\usage{
VariantFilters(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the filters
}
\description{
return the FILTERS for this variant
}
