% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantFlagAttribute}
\alias{VariantFlagAttribute}
\title{Get flag attribute from variant INFO}
\usage{
VariantFlagAttribute(vc, att)
}
\arguments{
\item{vc}{the variant}

\item{att}{the INFO/Attribute}
}
\value{
the the INFO attribute for the given key.
}
\description{
Get flag attribute from variant INFO
}
