% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantFloatAttribute}
\alias{VariantFloatAttribute}
\title{Get float attribute from variant INFO}
\usage{
VariantFloatAttribute(vc, att)
}
\arguments{
\item{vc}{the variant}

\item{att}{the INFO/Attribute}
}
\value{
the the INFO attribute for the given key.
}
\description{
Get float attribute from variant INFO
}
