% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantFormatIds}
\alias{VariantFormatIds}
\title{Get FORMAT keys for variant}
\usage{
VariantFormatIds(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the list FORMATs for this variant
}
\description{
Get FORMAT keys for variant
}
