% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantGenotype}
\alias{VariantGenotype}
\title{Get genotype for variant}
\usage{
VariantGenotype(vc, nameOrIdx)
}
\arguments{
\item{vc}{the variant}

\item{nameOrIdx}{the sample name (slower) or the 1-based sample index}
}
\value{
the given genotype
}
\description{
return the genotype for a variant
}
