% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantGenotypes}
\alias{VariantGenotypes}
\title{Get all genotypes for variant}
\usage{
VariantGenotypes(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the number of samples/genotypes for this variant
}
\description{
Get all genotypes for variant
}
