% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantGenotypesAlleleCounts}
\alias{VariantGenotypesAlleleCounts}
\title{Get allele counts per genotype}
\usage{
VariantGenotypesAlleleCounts(vc, allele_index = 1)
}
\arguments{
\item{vc}{the variant}

\item{allele_index}{The index of the allele to count (0-based)}
}
\value{
an integer vector of length \code{n_samples}
}
\description{
Counts the allele-counts per genotype for a given allele
}
\details{
The returned vector contains the number of occurance of a specific allele
in all genotypes.

For example, for a 3 sample VCF on a diploid variant with genotypes

\if{html}{\out{<div class="sourceCode">}}\preformatted{  GT1  GT2  GT3
  0/1  0/0  1/1
}\if{html}{\out{</div>}}

the result will be

\if{html}{\out{<div class="sourceCode">}}\preformatted{VariantGenotypesAlleleCounts(vc, 0) => c(1,2,0)
VariantGenotypesAlleleCounts(vc, 1) => c(1,0,2)
}\if{html}{\out{</div>}}
}
