% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantGenotypesAlleleIdx0}
\alias{VariantGenotypesAlleleIdx0}
\title{Get 0-based allele indexes for all genotypes}
\usage{
VariantGenotypesAlleleIdx0(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
an integer vector of length \code{ploidy * n_samples}
}
\description{
return the 0-based allele indizes for all genotypes of the variant
}
\details{
The returned vector contains the allele-indizes for all genotypes in sample order

For example, for a 3 sample VCF on a diploid variant, the returned vector could be

\if{html}{\out{<div class="sourceCode">}}\preformatted{  GT1  GT2  GT3
c(0,1, 0,0, 1,1)
}\if{html}{\out{</div>}}
}
