% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantGenotypesAlleleStrings}
\alias{VariantGenotypesAlleleStrings}
\title{Get genotype strings for all genotypes}
\usage{
VariantGenotypesAlleleStrings(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
a character vector of length \code{n_samples}
}
\description{
return the genotype strings for all genotypes of the variant
}
\details{
The returned vector contains the genotype strings in sample order.

For example, for a 3 sample VCF on a diploid variant, the returned vector could be

\if{html}{\out{<div class="sourceCode">}}\preformatted{   GT1    GT2    GT3
c("0/1", "0/0", "1/1")
}\if{html}{\out{</div>}}
}
