% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantGenotypesFlagAttribute}
\alias{VariantGenotypesFlagAttribute}
\title{Get FORMAT flag value on all genotypes}
\usage{
VariantGenotypesFlagAttribute(vc, att)
}
\arguments{
\item{vc}{the variant}
}
\value{
vector of logicals containing attribute values for all genotypes
}
\description{
Return a specific FORMAT flag value on all genotypes
}
\details{
The returned vector contains the flags by sample and attribute number (if the attribute comprises multiple flags).

For example, for a 3 sample VCF with a flag having a single logical, the returned vector could be

\if{html}{\out{<div class="sourceCode">}}\preformatted{   GT1    GT2   GT3
c(TRUE, FALSE, TRUE)
}\if{html}{\out{</div>}}
}
\seealso{
\link{VariantGenotypesIntAttribute},
\link{VariantGenotypesFloatAttribute}
}
