% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantGenotypesFloatAttribute}
\alias{VariantGenotypesFloatAttribute}
\title{Get FORMAT numeric value on all genotypes}
\usage{
VariantGenotypesFloatAttribute(vc, att)
}
\arguments{
\item{vc}{the variant}

\item{att}{the attribute name to retrieve}
}
\value{
vector of numerics containing attribute values for all genotypes
}
\description{
Return a specific FORMAT numeric value on all genotypes
}
\details{
The returned vector contains the values by sample and attribute number (if the attribute comprises multiple integer values).

For example, for a 3 sample VCF extracting the allelic read fraction (AF) on a multi-allelic variant, the results could look like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  GT1-ALT1, GT1-ALT2, GT2-ALT1, GT2-ALT2, GT3-ALT1, GT3-ALT2
c(     0.9,      0.1,      0.6,      0.1,      0.1,      0.4)
}\if{html}{\out{</div>}}
}
\seealso{
\link{VariantGenotypesFlagAttribute},
\link{VariantGenotypesIntAttribute}
}
