% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantGenotypesIntAttribute}
\alias{VariantGenotypesIntAttribute}
\title{Get FORMAT integer value on all genotypes}
\usage{
VariantGenotypesIntAttribute(vc, att)
}
\arguments{
\item{vc}{the variant}
}
\value{
vector of numerics containing attribute values for all genotypes
}
\description{
Return a specific FORMAT integer value on all genotypes
}
\details{
The returned vector contains the values by sample and attribute number (if the attribute comprises multiple integer values).

For example, for a 3 sample VCF extracting the allelic read depth (AD) on a singleton variant, the results could look like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  GT1-REF, GT1-ALT, GT2-REF, GT2-ALT, GT3-REF, GT3-ALT
c(     10,     100,      25,      94,      45,       7)
}\if{html}{\out{</div>}}
}
\seealso{
\link{VariantGenotypesFlagAttribute},
\link{VariantGenotypesFloatAttribute}
}
