% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantHasAttribute}
\alias{VariantHasAttribute}
\title{Check if variant has INFO attribute}
\usage{
VariantHasAttribute(vc, att)
}
\arguments{
\item{vc}{the variant}

\item{att}{the INFO/Attribute}
}
\value{
true if variant has INFO attribute
}
\description{
Check if variant has INFO attribute
}
