% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantHasFilter}
\alias{VariantHasFilter}
\title{Check if variant has specific filter}
\usage{
VariantHasFilter(vc, fn)
}
\arguments{
\item{vc}{the variant}

\item{fn}{filter name}
}
\value{
true if variant is filtered with 'fn'
}
\description{
Check if variant has specific filter
}
