% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantHasId}
\alias{VariantHasId}
\title{Check if variant has ID}
\usage{
VariantHasId(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the TRUE if variant have got an ID
}
\description{
return whether variant have got an ID
}
