% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantHasQual}
\alias{VariantHasQual}
\title{Check if QUAL is available for variant}
\usage{
VariantHasQual(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
QUAL is available for this variant
}
\description{
test if QUAL is available for this variant
}
