% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantInfoIds}
\alias{VariantInfoIds}
\title{Get INFO keys for variant}
\usage{
VariantInfoIds(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the list INFOs for this variant
}
\description{
Get INFO keys for variant
}
