% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantIntAttribute}
\alias{VariantIntAttribute}
\title{Get integer attribute from variant INFO}
\usage{
VariantIntAttribute(vc, att)
}
\arguments{
\item{vc}{the variant}

\item{att}{the INFO/Attribute}
}
\value{
the the INFO attribute for the given key.
}
\description{
Get integer attribute from variant INFO
}
