% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantIsFiltered}
\alias{VariantIsFiltered}
\title{Check if variant is filtered}
\usage{
VariantIsFiltered(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
variant is filtered
}
\description{
test is variant is filtered
}
