% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantIsSnp}
\alias{VariantIsSnp}
\title{Check if variant is SNP}
\usage{
VariantIsSnp(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
true if the variant is a SNP
}
\description{
return true if variant is a SNP
}
