% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantMaxPloidy}
\alias{VariantMaxPloidy}
\title{Get max ploidy for variant}
\usage{
VariantMaxPloidy(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
max ploidy
}
\description{
max ploidy for this variant
}
