% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantNAlleles}
\alias{VariantNAlleles}
\title{Get number of alleles for variant}
\usage{
VariantNAlleles(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the number of alleles for this variant
}
\description{
get the number of alleles for this variant
}
