% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantNSamples}
\alias{VariantNSamples}
\title{Get number of samples for variant}
\usage{
VariantNSamples(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the number of samples/genotypes for this variant
}
\description{
Get number of samples for variant
}
