% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantPos}
\alias{VariantPos}
\title{Get variant position}
\usage{
VariantPos(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the starting position for this variant
}
\description{
get chromosome POS for this variant
}
