% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantReference}
\alias{VariantReference}
\title{Get reference allele for variant}
\usage{
VariantReference(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the reference allele for this variant
}
\description{
get the reference allele for this variant
}
