% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantSnpEff}
\alias{VariantSnpEff}
\title{Get SnpEff annotation table for variant}
\usage{
VariantSnpEff(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
SNPEFF table for this variant or NULL
}
\description{
Get SnpEff annotation table for variant
}
