% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantStart}
\alias{VariantStart}
\title{Get variant start position (alias)}
\usage{
VariantStart(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the starting position for this variant
}
\description{
alias of VariantPos
}
