% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantStop}
\alias{VariantStop}
\title{Get variant stop position (alias)}
\usage{
VariantStop(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the END position for this variant
}
\description{
alias of VariantEnd
}
