% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantStringAttribute}
\alias{VariantStringAttribute}
\title{Get string attribute from variant INFO}
\usage{
VariantStringAttribute(vc, att, split = TRUE)
}
\arguments{
\item{vc}{the variant}

\item{att}{the INFO/Attribute}

\item{split}{split strings using commas}
}
\value{
the the INFO attribute for the given key.
}
\description{
Get string attribute from variant INFO
}
