% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantTid}
\alias{VariantTid}
\title{Get chromosome numeric index (tid)}
\usage{
VariantTid(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the numeric index for this variant
}
\description{
get the numeric index (tid) of the chromosome for this variant
}
