% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantTypes}
\alias{VariantTypes}
\title{Get types for variant}
\usage{
VariantTypes(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
the types for this variant
}
\description{
return the types for this variant (as defined in htslib)
}
