% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantVep}
\alias{VariantVep}
\title{Get VEP annotation table for variant}
\usage{
VariantVep(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
VEP table for this variant or NULL
}
\description{
Get VEP annotation table for variant
}
