% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIExtractRanges}
\alias{VBIExtractRanges}
\title{Extract variant ranges from a VBI index pointer}
\usage{
VBIExtractRanges(VbiPtr, n = NA)
}
\arguments{
\item{VbiPtr}{External pointer to VBI index}

\item{n}{Number of ranges to extract (default: all)}
}
\value{
data.frame with chrom, start, end, label
}
\description{
Extract variant ranges from a VBI index pointer
}
