% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIFilters}
\alias{VBIFilters}
\title{Get FILTER fields table from VBI VCF context}
\usage{
VBIFilters(vbi_vcf_ctx)
}
\arguments{
\item{vbi_vcf_ctx}{VBI VCF context object from VCFLoad()}
}
\value{
data.frame with FILTER field definitions
}
\description{
Get FILTER fields table from VBI VCF context
}
