% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIFormats}
\alias{VBIFormats}
\title{Get FORMAT fields table from VBI VCF context}
\usage{
VBIFormats(vbi_vcf_ctx)
}
\arguments{
\item{vbi_vcf_ctx}{VBI VCF context object from VCFLoad()}
}
\value{
data.frame with FORMAT field definitions
}
\description{
Get FORMAT fields table from VBI VCF context
}
