% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIIndex}
\alias{VBIIndex}
\title{VBI Index}
\usage{
VBIIndex(VcfPath, VbiPath, Threads = 1)
}
\arguments{
\item{VcfPath}{Path to the VCF/BCF file}

\item{VbiPath}{Path to the VBI index file}

\item{Threads}{Number of threads to use (default: 1)}
}
\value{
Path to index file
}
\description{
Create a VBI index for a VCF/BCF file.
}
