% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIIndexLoad}
\alias{VBIIndexLoad}
\title{Load a VBI index file and return an external pointer}
\usage{
VBIIndexLoad(vbi_path)
}
\arguments{
\item{vbi_path}{Path to the VBI index file}
}
\value{
External pointer to the loaded VBI index
}
\description{
Load a VBI index file and return an external pointer
}
