% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIIndexMemoryUsage}
\alias{VBIIndexMemoryUsage}
\title{Get C-level memory usage of a VBI index pointer (in bytes)}
\usage{
VBIIndexMemoryUsage(vbi_ptr)
}
\arguments{
\item{vbi_ptr}{External pointer to VBI index}
}
\value{
Numeric, memory usage in bytes
}
\description{
Get C-level memory usage of a VBI index pointer (in bytes)
}
