% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIInfos}
\alias{VBIInfos}
\title{Get INFO fields table from VBI VCF context}
\usage{
VBIInfos(vbi_vcf_ctx)
}
\arguments{
\item{vbi_vcf_ctx}{VBI VCF context object from VCFLoad()}
}
\value{
data.frame with INFO field definitions
}
\description{
Get INFO fields table from VBI VCF context
}
