% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBINSamples}
\alias{VBINSamples}
\title{Get number of samples in VBI VCF context}
\usage{
VBINSamples(vbi_vcf_ctx)
}
\arguments{
\item{vbi_vcf_ctx}{VBI VCF context object from VCFLoad()}
}
\value{
Integer, number of samples
}
\description{
Get number of samples in VBI VCF context
}
