% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIPrintHeaderMetadata}
\alias{VBIPrintHeaderMetadata}
\title{Print VCF/BCF header metadata from a VBI context}
\usage{
VBIPrintHeaderMetadata(vbi_vcf_ctx)
}
\arguments{
\item{vbi_vcf_ctx}{VBI VCF context object from VCFLoad()}
}
\value{
Invisible NULL
}
\description{
Prints a summary of the header fields and sample names for the file from a VBI VCF context object.
}
