% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIPrintIndex}
\alias{VBIPrintIndex}
\title{Print the first n lines of a VBI index file}
\usage{
VBIPrintIndex(vbi_path, n = 10)
}
\arguments{
\item{vbi_path}{Path to the VBI index file}

\item{n}{Number of lines to print (default: 10)}
}
\description{
Print the first n lines of a VBI index file
}
