% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIQueryRange}
\alias{VBIQueryRange}
\title{Query VBI index by region}
\usage{
VBIQueryRange(VcfPath, VbiPath, Region, Threads = 1)
}
\arguments{
\item{VcfPath}{Path to the VCF/BCF file}

\item{VbiPath}{Path to the VBI index file}

\item{Region}{Region string (e.g., "chr1:1000-2000") for query}

\item{Threads}{Number of threads to use (default: 1)}
}
\value{
List of results
}
\description{
Query VBI index by region
}
