% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBISample2Index}
\alias{VBISample2Index}
\title{Convert sample name to index in VBI VCF context}
\usage{
VBISample2Index(vbi_vcf_ctx, sample_name)
}
\arguments{
\item{vbi_vcf_ctx}{VBI VCF context object from VCFLoad()}

\item{sample_name}{Character, sample name}
}
\value{
Integer, 1-based sample index (0 if not found)
}
\description{
Convert sample name to index in VBI VCF context
}
