% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBISampleAt}
\alias{VBISampleAt}
\title{Get sample at specific index from VBI VCF context}
\usage{
VBISampleAt(vbi_vcf_ctx, index)
}
\arguments{
\item{vbi_vcf_ctx}{VBI VCF context object from VCFLoad()}

\item{index}{1-based sample index}
}
\value{
Character, sample name
}
\description{
Get sample at specific index from VBI VCF context
}
