% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBISamples}
\alias{VBISamples}
\title{Get VBI VCF context samples}
\usage{
VBISamples(vbi_vcf_ctx)
}
\arguments{
\item{vbi_vcf_ctx}{VBI VCF context object from VCFLoad()}
}
\value{
Character vector of sample names
}
\description{
Retrieve sample names from a VBI VCF context object.
}
