% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VCFLoad}
\alias{VCFLoad}
\title{Load a VCF file with VBI index integration}
\usage{
VCFLoad(vcf_path, vbi_path = NULL)
}
\arguments{
\item{vcf_path}{Path to the VCF/BCF file}

\item{vbi_path}{Path to the VBI index file (optional, will auto-detect if NULL)}
}
\value{
External pointer to VBI VCF context object
}
\description{
Creates a unified VCF object that includes the VCF file, header information,
and VBI index for fast queries. This follows the RBCF pattern but with VBI integration.
}
\examples{
\dontrun{
vcf_file <- system.file("exdata", "imputed.gt.vcf.gz", package = "RBCFLib")
vcf_obj <- VCFLoad(vcf_file)

# Query with basic fields
hits <- VBIQueryRegion(vcf_obj, "chr21:5030082-5030356")

# Query with INFO fields
hits_with_info <- VBIQueryRegion(vcf_obj, "chr21:5030082-5030356",
                                 include_info = TRUE)
}
}
