% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcf.R
\name{VariantTypes}
\alias{VariantTypes}
\title{Get types for variant}
\usage{
VariantTypes(vc)
}
\arguments{
\item{vc}{the variant}
}
\value{
character vector of variant types present in this variant
}
\description{
Return the types for this variant (as defined in htslib)
}
\details{
This function returns a character vector of variant types present in the record.
Possible types include: "REF", "SNP", "MNP", "INDEL", "INS", "DEL", "BND",
"OVERLAP", and "OTHER". Modern HTSlib supports structural variants through
"BND" (breakends) and complex variants.
}
