% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIPrintIndex}
\alias{VBIPrintIndex}
\title{Print the first n lines of a VBI index (from an external pointer)}
\usage{
VBIPrintIndex(vbi_ptr, n = 10)
}
\arguments{
\item{vbi_ptr}{External pointer to loaded VBI index}

\item{n}{Number of lines to print (default: 10)}
}
\description{
Print the first n lines of a VBI index (from an external pointer)
}
