% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBI.R
\name{VBIQueryByIndices}
\alias{VBIQueryByIndices}
\title{Query VBI index by marker index range}
\usage{
VBIQueryByIndices(VcfPath, VbiPath, StartingIndice, EndingIndice, Threads = 1)
}
\arguments{
\item{VcfPath}{Path to the VCF/BCF file}

\item{VbiPath}{Path to the VBI index file}

\item{StartingIndice}{Start index for index-based query}

\item{EndingIndice}{End index for index-based query}

\item{Threads}{Number of threads to use (default: 1)}
}
\value{
List of results
}
\description{
Query VBI index by marker index range
}
